+++ boot-high-rps-sample/Dockerfile
@@ -4,6 +4,7 @@ COPY ${JAR_FILE} app.jar
 
 # Run as non-root user
 RUN addgroup --system spring && adduser --system --ingroup spring spring
+RUN apt-get update && apt-get install -y curl && rm -rf /var/lib/apt/lists/*
 USER spring:spring
 
 EXPOSE 8080
@@ -18,4 +19,4 @@ ENTRYPOINT ["java", \
 
 # Health check for container orchestration
 HEALTHCHECK --interval=30s --timeout=3s --start-period=40s \
-  CMD curl -f http://localhost:8080/actuator/health || exit 1
+  CMD java -cp /app.jar org.springframework.boot.loader.tools.HealthCheckClient http://localhost:8080/actuator/health || exit 1
