+++ boot-high-rps-sample/src/main/java/com/example/highrps/repository/StatsEntity.java
@@ -3,8 +3,10 @@ package com.example.highrps.repository;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 @Entity
+@Table(name = "stats_entity")
 public class StatsEntity {
 
     @Id
