+++ boot-high-rps-sample/src/main/resources/application.properties
@@ -46,3 +46,8 @@ app.batch.delay-ms=5000
 
 # JVM hints (informational only) - tune as required for performance
 # Recommended JVM flags in production: -XX:+UseZGC -Xms1g -Xmx2g -XX:+UseCompressedOops
+
+app.kafka.events-topic.partitions=3
+app.kafka.events-topic.replication-factor=1
+app.kafka.stats-aggregates-topic.partitions=3
+app.kafka.stats-aggregates-topic.replication-factor=1
